+++ backend/server/src/main/java/com/airbnb/controller/PaymentDetailsController.java
@@ -1,9 +1,6 @@
 package com.airbnb.controller;
 
-import com.airbnb.dto.mainpage.MainPageResponse;
 import com.airbnb.dto.payment.PaymentDetailsDto;
-import com.airbnb.dto.room.RoomListResponse;
-import com.airbnb.service.AirbnbService;
 import com.airbnb.service.PaymentDetailsService;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -11,23 +8,11 @@ import org.springframework.web.bind.annotation.RestController;
 
 @RestController
 @RequestMapping("/api")
-public class AirbnbController {
-    private final AirbnbService airbnbService;
+public class PaymentDetailsController {
     private final PaymentDetailsService paymentDetailsService;
 
-    public AirbnbController(AirbnbService airbnbService, PaymentDetailsService paymentDetailsService) {
-        this.airbnbService = airbnbService;
+    public PaymentDetailsController(PaymentDetailsService paymentDetailsService) {
         this.paymentDetailsService = paymentDetailsService;
-    }
-
-    @GetMapping("/mainPage")
-    public MainPageResponse getMainPage() {
-        return airbnbService.getMainPage();
-    }
-
-    @GetMapping("/search")
-    public RoomListResponse getRoomList() {
-        return airbnbService.getRoomList();
     }
 
     @GetMapping("/popup")
