+++ backend/server/src/main/java/com/airbnb/domain/auth/Token.java
@@ -1,4 +1,4 @@
-package com.airbnb.domain;
+package com.airbnb.domain.auth;
 
 public class Token {
     private final String accessToken;
