+++ backend/server/src/main/java/com/airbnb/repository/UserRepository.java
@@ -1,6 +1,6 @@
 package com.airbnb.repository;
 
-import com.airbnb.domain.User;
+import com.airbnb.domain.auth.User;
 import com.airbnb.exception.SqlException;
 import org.springframework.stereotype.Repository;
