+++ backend/server/src/main/java/com/airbnb/service/AuthService.java
@@ -1,6 +1,6 @@
 package com.airbnb.service;
 
-import com.airbnb.domain.Auth;
+import com.airbnb.domain.auth.Auth;
 import com.airbnb.dto.AccessTokenResponse;
 import com.airbnb.dto.UserDto;
 import com.airbnb.exception.AuthenticationException;
