+++ backend/server/src/main/java/com/airbnb/service/RoomListService.java
@@ -1,0 +1,18 @@
+package com.airbnb.service;
+
+import com.airbnb.dto.room.RoomListResponse;
+import com.airbnb.repository.RoomListRepository;
+import org.springframework.stereotype.Service;
+
+@Service
+public class RoomListService {
+    private final RoomListRepository roomListRepository;
+
+    public RoomListService(RoomListRepository roomListRepository) {
+        this.roomListRepository = roomListRepository;
+    }
+    public RoomListResponse getRoomList() {
+        return RoomListResponse.from(roomListRepository.getRoomList());
+    }
+
+}
